DUNE/Parsers: Moved HDF5Reader to core.
