+++ src/Simulators/StreamSpeed/ModelDataStreamGenerator.hpp
@@ -33,8 +33,9 @@
 #include <string>
 #include <vector>
 
+#include "DUNE/Parsers/HDF5Reader.hpp"
+
 #include "Grid.hpp"
-#include "HDF5Reader.hpp"
 #include "StreamGenerator.hpp"
 
 namespace Simulators
@@ -91,7 +92,7 @@ namespace Simulators
                  double time = 0.0) const override;
 
       private:
-        HDF5Reader m_file;
+        DUNE::Parsers::HDF5Reader m_file;
         //! Velocity in the East direction.
         std::vector<double> m_u;
         //! Velocity in the North direction.
